reorganize so logic can be reused in resourceconfig
