+++ plugins/codeamp/graphql/helpers.go
@@ -7,7 +7,6 @@ import (
 	"github.com/codeamp/circuit/plugins"
 	"github.com/codeamp/circuit/plugins/codeamp/model"
 	"github.com/codeamp/transistor"
-	"github.com/jinzhu/gorm"
 )
 
 func AppendPluginService(pluginServices []plugins.Service, service model.Service, spec model.ServiceSpec) []plugins.Service {
@@ -131,14 +130,4 @@ func BuildReleasePayload(release model.Release, project model.Project, environme
 		Services:   services,
 		IsRollback: release.IsRollback,
 	}
-}
-
-func GetProjectExtensionsWithRoute53Subdomain(subdomain string, db *gorm.DB) []model.ProjectExtension {
-	var existingProjectExtensions []model.ProjectExtension
-
-	if db.Where("custom_config ->> 'subdomain' ilike ?", subdomain).Find(&existingProjectExtensions).RecordNotFound() {
-		return []model.ProjectExtension{}
-	}
-
-	return existingProjectExtensions
 }
